+++ apps/site/src/pages/account/api-keys-panel/api-keys-context.tsx
@@ -1,27 +1,6 @@
-import { createContext, Dispatch, SetStateAction, useContext } from "react";
+import { createContext, useContext } from "react";
 
-import { UserFacingApiKeyProperties } from "../../../lib/api/model/api-key.model";
-
-export type KeyAction = "rename" | "revoke";
-
-export type KeyActionStatus =
-  | {
-      publicId: string;
-      action: KeyAction;
-    }
-  | undefined;
-
-export interface ApiKeysContextValue {
-  apiKeys: UserFacingApiKeyProperties[];
-  setApiKeys: Dispatch<SetStateAction<UserFacingApiKeyProperties[]>>;
-  isCreatingNewKey: boolean;
-  setIsCreatingNewKey: Dispatch<SetStateAction<boolean>>;
-  newlyCreatedKeyIds: string[];
-  setNewlyCreatedKeyIds: Dispatch<SetStateAction<string[]>>;
-  keyActionStatus: KeyActionStatus;
-  setKeyActionStatus: Dispatch<SetStateAction<KeyActionStatus>>;
-  fetchAndSetApiKeys: () => Promise<void>;
-}
+import { ApiKeysContextValue } from "./types";
 
 export const ApiKeysContext = createContext<ApiKeysContextValue | null>(null);
